Move integration tests
The remaining tests in the lib.rs file appear to be integration tests
for example files.

According to [the book](https://doc.rust-lang.org/stable/book/ch11-03-test-organization.html#integration-tests) these should go into `tests/` and use the crate externally.

This commit makes this change and necessary exposure
of public methods.
