+++ src/ideal.rs
@@ -92,7 +92,7 @@ impl Ideal {
         Ideal(vec![val; dimension])
     }
 
-    pub(crate) fn from_vec(vec: Vec<Coef>) -> Ideal {
+    pub fn from_vec(vec: Vec<Coef>) -> Ideal {
         Ideal(vec)
     }
 
@@ -100,19 +100,19 @@ impl Ideal {
         self.0.iter().enumerate().all(|(i, &x)| x <= other.0[i])
     }
 
-    pub(crate) fn len(&self) -> usize {
+    pub fn len(&self) -> usize {
         self.0.len()
     }
 
-    pub(crate) fn get(&self, i: usize) -> Coef {
+    pub fn get(&self, i: usize) -> Coef {
         self.0[i]
     }
 
-    pub(crate) fn set(&mut self, state: usize, val: Coef) {
+    pub fn set(&mut self, state: usize, val: Coef) {
         self.0[state] = val;
     }
 
-    pub(crate) fn intersection(x: &Ideal, ideal: &Ideal) -> Ideal {
+    pub fn intersection(x: &Ideal, ideal: &Ideal) -> Ideal {
         debug_assert_eq!(x.len(), ideal.len());
         Ideal(
             x.0.iter()
@@ -124,7 +124,7 @@ impl Ideal {
     }
 
     #[allow(dead_code)]
-    pub(crate) fn from_non_zero_coefs(
+    pub fn from_non_zero_coefs(
         dim: usize,
         partition: &[coef],
         predecessors: &[usize],
@@ -137,11 +137,11 @@ impl Ideal {
         Ideal(result)
     }
 
-    pub(crate) fn all_omega(&self, succ: &[usize]) -> bool {
+    pub fn all_omega(&self, succ: &[usize]) -> bool {
         succ.iter().all(|&i| self.get(i) == OMEGA)
     }
 
-    pub(crate) fn round_up(&mut self, max_finite_value: coef) -> Ideal {
+    pub fn round_up(&mut self, max_finite_value: coef) -> Ideal {
         Ideal(
             self.0
                 .iter()
@@ -150,7 +150,7 @@ impl Ideal {
         )
     }
 
-    pub(crate) fn round_down(&mut self, upper_bound: coef, dim: usize) {
+    pub fn round_down(&mut self, upper_bound: coef, dim: usize) {
         for i in 0..dim {
             if let Coef::Value(x) = self.get(i) {
                 if x > upper_bound {
@@ -160,7 +160,7 @@ impl Ideal {
         }
     }
 
-    pub(crate) fn some_finite_coordinate_is_larger_than(&self, upper_bound: coef) -> bool {
+    pub fn some_finite_coordinate_is_larger_than(&self, upper_bound: coef) -> bool {
         self.0
             .iter()
             .any(|&x| x < OMEGA && x > Coef::Value(upper_bound))
@@ -178,7 +178,7 @@ impl Ideal {
         content
     }
 
-    pub(crate) fn iter(&self) -> impl Iterator<Item = &Coef> {
+    pub fn iter(&self) -> impl Iterator<Item = &Coef> {
         self.0.iter()
     }
 
@@ -190,7 +190,7 @@ impl Ideal {
         }
     }
 
-    pub(crate) fn clone_and_decrease(&self, i: usize, maximal_finite_value: coef) -> Ideal {
+    pub fn clone_and_decrease(&self, i: usize, maximal_finite_value: coef) -> Ideal {
         let mut result: Ideal = self.clone();
         let c = result.0[i];
         debug_assert!(c != Coef::Value(0));
